Add initial connectivity to SmartThings.
Handles the first two communication tasks:
- Registers the client uri for device updates
- Retrieves initial device definitions/statuses

Also, had to change things around so the settings are instantiated within the Vue app. This is just easier. This means we could remove the IPC communication for the settings (which, btw, did not transfer the class, but rather, just the object data).

Also, had to tell browser winder to have web security off so that CORS wouldn't complain about the SmartThings calls.