Add Daily Scripture screen.
This is a sample of a screen with completely different content and behavior. It loads the Daily Scripture from WOL and display it on a screen.

This required an improvement to Dashboard to load the definitions of the component into the component via v-model.  This will be used by future screens/components.