Improve way speaking executes to improve smoothness and allow stop.
- All speech is downloaded up front, THEN spoken. This allows for smoother speech because there is no brief delay when downloading next snippet.
- Speech sessions are tracked and a token is returned. The stop() method can then be passed a speech session token to stop existing speech.
- The speak() method can receive a callback so that action can be taken when speech finishes.

All of these changes help the user of the library have better control over the start/stop of speech.

To accomplish this, the token is generated by hashing the text and the current time. The token provided by the play-sound utility is stored in the `speaking` object under the key of the token. This can be used to stop the playing audio.

It would be nice to improve this to more cleanly have a sync and promise version. For now it's only one version that passes you the token and it executes a promise-based handleSpeak() method.